go.d snmp: add sys descr, contact and loc as host labels for vnode
